% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.rel.R
\name{CA.plotCI}
\alias{CA.plotCI}
\title{Plot 95 percent confidence interval plots for calibration statistics}
\usage{
CA.plotCI(CA.rel, ylim = list(C = NULL, OU = c(0), NRI = NULL))
}
\arguments{
\item{CA.rel}{Object created by the CA.rel function}

\item{ylim}{A list with the limitation of the y axis for each calibration statistic.}
}
\description{
\code{CA.plotCI()} takes the argument of \code{CA.rel()} and and plots the confidence
   interval plots for the C, O/U, and NRI calibration statistics.
}
\details{
For a more detailed explanation see the github page for a manual.
   \url{https://github.com/IngerMathilde/CArelationship}
}
\examples{
library(jtools)
data(metamemoryCA)

# Create a CI plot for the C, OU, NRI statistics for high and low metamemory raters of two
# different components.

ChMM <- CA.rel(data = metamemoryCA, confidence = "Confidence",
                        correct = "ChoiceCorrect", test = "CAL",
                        var = c("Rater.EMS.Relative.Face.Recognition",
                                "Rater.EMS.Eyewitness.Ability"),
                        var.levels = c('Low', 'High'),
                        confidenceLevels = list(c(0,20),c(30,40), c(50,60), c(70,80), c(90,100)),
                        jack = TRUE)
CA.plotCI(ChMM)

}
\author{
Inger van Boeijen <\email{inger.vb.r@gmail.com}>
}
