% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.rel.R
\name{CA.print}
\alias{CA.print}
\title{Print overview table per confidence level}
\usage{
CA.print(CA.rel)
}
\arguments{
\item{CA.rel}{Object created by the CA.rel function}
}
\description{
\code{CA.print()} takes the argument of \code{CA.rel()} and prints the seperate
   calibration tables (mean, prop correct, etc.) for each confidence interval per variable
}
\details{
For a more detailed explanation see the github page for a manual.
   \url{https://github.com/IngerMathilde/CArelationship}
}
\examples{
library(jtools)
data(metamemoryCA)

# Create object for the whole dataset and print proportion correct per level of confidence

All <- CA.rel(data = metamemoryCA, confidence = "Confidence",
                       correct = "ChoiceCorrect", test = "CAL",
                       confidenceLevels = c(0,10,20,30,40,50,60,70,80,90,100))

CA.print(All)


}
\author{
Inger van Boeijen <\email{inger.vb.r@gmail.com}>
}
