% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.rel.R
\name{CA.table}
\alias{CA.table}
\title{Prints table for calibration statistics}
\usage{
CA.table(CA.rel)
}
\arguments{
\item{CA.rel}{Object created by the CA.rel function}
}
\description{
\code{CA.table()} takes the argument of \code{CA.rel()} and prints a table with the
   C, O/U and NRI stats
}
\details{
For a more detailed explanation see the github page for a manual.
   \url{https://github.com/IngerMathilde/CArelationship}
}
\examples{
data(metamemoryCA)

# Compare choosers vs. nonchoosers with collapsed confidence groups and Jackknife SE
Ch <- CA.rel(data = metamemoryCA, confidence = "Confidence",
                      correct = "ChoiceCorrect", test = "CAL", var = "ChoiceChooser",
                      confidenceLevels = list(c(0,20),c(30,40), c(50,60), c(70,80), c(90,100)),
                      jack = TRUE)

CA.table(Ch)

}
\author{
Inger van Boeijen <\email{inger.vb.r@gmail.com}>
}
