% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.rel.R
\docType{data}
\name{metamemoryCA}
\alias{metamemoryCA}
\title{Example dataset for confidence accuracy relationship}
\format{data frame with 356 rows and 9 variables.}
\usage{
data(metamemoryCA)
}
\description{
Dataset including identification accuracy, identification confidence estimates and
metamemory scores
}
\examples{
data(metamemoryCA)
Ch <- CA.rel(data = metamemoryCA, confidence = "Confidence",
                      correct = "ChoiceCorrect", test = "CAL", var = "ChoiceChooser",
                      confidenceLevels = list(c(0,20),c(30,40), c(50,60), c(70,80), c(90,100)),
                      jack = TRUE)

}
\keyword{datasets}
